+++ main/main.cpp
@@ -10,6 +10,7 @@
 #include "water_tank_app.hpp"
 #include "water_tank_nvs.hpp"
 #include "persistence_backend.hpp"
+#include "nvs_core.hpp"
 #include "secrets.hpp"
 #include "ota_manager.hpp"
 #include "button_ota_trigger.hpp"
@@ -116,8 +117,8 @@ static RtcBackend rtc_core_backend(&g_rtc_core, sizeof(CoreStorage));
 static NvsBackend nvs_core_backend{nvs_hal, CORE_NVS_KEY};
 static NvsCore nvs_core{rtc_core_backend, nvs_core_backend};
 
-static RTC_DATA_ATTR WaterTankStats g_rtc_tank;
-static RtcBackend rtc_stats_backend(&g_rtc_tank, sizeof(WaterTankStats));
+static RTC_DATA_ATTR WaterTankStorage g_rtc_tank;
+static RtcBackend rtc_stats_backend(&g_rtc_tank, sizeof(WaterTankStorage));
 static NvsBackend nvs_stats_backend{nvs_hal, STATS_NVS_KEY};
 static WaterTankNvs nvs_tank{rtc_stats_backend, nvs_stats_backend};